The first one.
You might be wondering why this commit has such a significant amount of
work. That's because this project was initially started in another
repository. I had to down the zip and transfer it under my name thus
this commit introduces a semi-functional Rails app.

The main idea behind this app is to allow Shopify merchants to keep
their store products in sync with their vendor's inventory. The app
accepts a CSV file which is then parsed (based on the columns you
specify), all new products are added to your store and the rest are
updated.

You do require to do some coding to get this app up and running however
the code is well documented and it should be fairly simple to follow
along!